Adds additional URL generation support.
This serializer may run outside of the normal request/response cycle,
for example when used in the 'cache_lakes' management command. In this
context the serializer will consequently have no request object.

In such cases, use the first available entry in 'ALLOWED_HOSTS' to
use as the hostname for the generated URL.