Adds new cloudformation defintions for 'vpc-cloud'.
The existing production environment resides in the original (legacy)
VPC 'vpc-prod' which is planned to be pruned of active projects.

Additionally, the EC2 instance is an AL2 host (which has since been
deprecated) and its RDS instance currently runs PostgreSQL 11 which
is EOL and will incur additional charges unless it is upgraded.

Due to the lack of interest/funding/collaboration, this project will
be moved to the cloud VPC in an "archived" configuration using legacy
services in an isolated docker environment until it is decided whether
the project will receive new development or will be retired.